Compile on certain 64 bit platforms.
[git-p4: depot-paths = "//depot/qt/main/": change = 166189]

+++ src/gui/kernel/qapplication.cpp
@@ -2601,10 +2601,10 @@ bool QApplication::notify(QObject *receiver, QEvent *e)
                "QApplication::sendEvent",
                QString::fromLatin1("Cannot send events to objects owned by a different thread "
                                    "(%1). Receiver '%2' (of type '%3') was created in thread %4")
-               .arg(QString::number((int)QThread::currentThread(), 16))
+               .arg(QString::number((qlonglong)(QThread::currentThread()), 16))
                .arg(receiver->objectName())
                .arg(receiver->metaObject()->className())
-               .arg(QString::number((int)receiver->thread(), 16))
+               .arg(QString::number((qlonglong)(receiver->thread())), 16)
                .toLatin1().constData());
 
 #ifdef QT_COMPAT
